implemented exception handling of springboot

+++ journal-api/src/main/java/com/shubham/journal_api/controller/TradeController.java
@@ -1,8 +1,12 @@
 package com.shubham.journal_api.controller;
 
+import com.shubham.journal_api.exception.TradeNotFoundException;
 import com.shubham.journal_api.model.Trade;
 import com.shubham.journal_api.repository.TradeRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,7 +21,7 @@ public class TradeController{
     private TradeRepository tradeRepository;
 
     //temporary storage
-    private List<Trade> trades = new ArrayList<>();
+   // private List<Trade> trades = new ArrayList<>();
 
     //Get all trades
     @GetMapping
@@ -28,34 +32,36 @@ public class TradeController{
     // GET single trade by ID
     @GetMapping("/{id}")
     public Trade getTradeById(@PathVariable Long id) {
-        return tradeRepository.findById(id).orElse(null);
+        return tradeRepository.findById(id).orElseThrow(() -> new TradeNotFoundException(id));
     }
 
     //add new trade
     @PostMapping
-    public Trade addNewTrade(@RequestBody Trade trade){
-        return tradeRepository.save(trade);
+    public ResponseEntity<Trade> addNewTrade(@Valid @RequestBody Trade trade){
+        Trade savedTrade = tradeRepository.save(trade);
+        return new ResponseEntity<>(savedTrade, HttpStatus.CREATED);  //201 status will be there
+
 
     }
 
     //counting trades number
-    @GetMapping("/count")
-    public String getTradeCount(){
-        return "Total trades: " + trades.size();
-
-    }
-
-    //trade pnl
-    @GetMapping("/{index}/pnl")
-    public String getTradePnL(@PathVariable int index){
-        if(index >= 0 && index < trades.size()) {
-            Trade trade = trades.get(index);
-            double pnl = trade.calculatePnL();
-            return String.format("P&L for %s: ₹%.2f %s",
-                    trade.getSymbol(), pnl, pnl > 0 ? "✓" : "✗");
-        }
-        return "Trade not found" ;
-    }
+//    @GetMapping("/count")
+//    public String getTradeCount(){
+//        return "Total trades: " + trade.size();
+//
+//    }
+//
+//    //trade pnl
+//    @GetMapping("/{index}/pnl")
+//    public String getTradePnL(@PathVariable int index){
+//        if(index >= 0 && index < trade.size()) {
+//            Trade trade = trade.get(index);
+//            double pnl = trade.calculatePnL();
+//            return String.format("P&L for %s: ₹%.2f %s",
+//                    trade.getSymbol(), pnl, pnl > 0 ? "✓" : "✗");
+//        }
+//        return "Trade not found" ;
+//    }
 
     //total pnl
     @GetMapping("/stats/total-pnl")
@@ -85,25 +91,28 @@ public class TradeController{
 
     //put - update trade
     @PutMapping("/{id}")
-    public Trade updateTrade(@PathVariable Long id, @RequestBody Trade tradeDetails){
-        Trade trade = tradeRepository.findById(id).orElse(null);
-        if(trade != null){
+    public Trade updateTrade(@PathVariable Long id, @Valid @RequestBody Trade tradeDetails){
+        Trade trade = tradeRepository.findById(id).orElseThrow(() ->new TradeNotFoundException(id));
+
             trade.setSymbol(tradeDetails.getSymbol());
             trade.setType(tradeDetails.getType());
             trade.setEntryPrice(tradeDetails.getEntryPrice());
             trade.setExitPrice(tradeDetails.getExitPrice());
             trade.setQuantity(tradeDetails.getQuantity());
             trade.setStrategy(tradeDetails.getStrategy());
+
             return tradeRepository.save(trade);
-        }
-        return null;
+
     }
 
     //Delete trade
     @DeleteMapping("/{id}")
-    public String deleteTrade(@PathVariable Long id){
+    public ResponseEntity<String> deleteTrade(@PathVariable Long id){
+        if (!tradeRepository.existsById(id)){
+            throw new TradeNotFoundException(id);
+        }
         tradeRepository.deleteById(id);
-        return "Trade deleted with id "+ id;
+        return ResponseEntity.ok("Trade deleted with id: " + id);
 
     }
